Refactor data source config
